+++ src/snakeoil/__init__.py
@@ -13,6 +13,6 @@ This library is a bit of a grabbag of the following:
 __title__ = "snakeoil"
 # TODO: wire these all against py_build rendered values, or against git.  Don't hardcode it here.
 # Once that's done, extend it to the rest of the ecosystem.
-__version__ = "0.11.6.dev0"
-__version_info__ = (0, 11, 6, "dev0")
+__version__ = "0.11.6"
+__version_info__ = (0, 11, 6)
 __python_mininum_version__ = (3, 12, 0)